Adding Dockerfile - first working version
This commit adds the first working version of a Dockerfile that contains the entire motion package.
It also edits the motion.conf file to my existing settings and then allows use of build arguements to define a username and password for webaccess to the stream.
Installs the no-ip toolset so that I can access the stream from anywhere using a domain name